pyproject.toml
[project]
name = "mwaa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apache-airflow>=2.10.5",
    "structlog>=25.2.0",
]

[dependency-groups]
dev = [
    "apache-airflow>=2.10.5",
    "pytest>=8.3.5",
]
